rest: api for accepting an un-managed node
Enhanced the below api to accept an un-managed node
    - PUT "/api/<ver>/nodes"

If only arguments passed to the PUT request is hostname
and ssh fingerprint, its treated as accept node and
accordingly accept is invoked.

In case all the required inputs like hostname, ssh fingerprint
, user and password is passed, its treated as addition
of an adhoc node and accordingly add and accept is
invoked.

Signed-off-by: Shubhendu Tripathi <shtripat@redhat.com>