Eventing: This patch adds basic eventing capability to SkyRing
In this patch, have modified the push_events.sls reactor file to
push the node dbus events to Skyring by making an rpc call. In
RPC Method the events are handled appropriately using go routines
to route and invoke respective event handlers. As of
this patch only handling done is writing the events in DB
under "node_events" collection. Also this patch adds a basic API
to retrieve list of all the events (/api/v1/events).

Change-Id: I6e73224157219a80fd21636d90b6a48ec61dc3ea
Signed-off-by: Darshan N <darshan.n.2024@gmail.com>