saltwrapper: Delete the minion key while ignore
Modified to delete the minion key as part of ignore
node.

Reject key is bit risky as later to accept the
minion back, keys need to be deleted in master and
minion and then re-start the minion.

In case of delete, only minions re-start would suffice.

NOTE: This ignore node would be invoked only in case of
forget cluster from skyring.

Change-Id: Ic09db6d0da77fcc69e3df17e3c50e930070915eb
Signed-off-by: Shubhendu Tripathi <shtripat@redhat.com>