websocket integration to push events to UI
As part skyring startup a master sync would sync all the entities for the cluster with their changes in status etc.
Post start, the state change and any alert/notifications would be percolated to UI using the websocket push for the events to UI.

Websocket starts on port 8081. Make sure the port is open and available.

Change-Id: I2af5026278394ae8b722b2fa505696d3482e21d9
Signed-off-by: kamlesh <kaverma@redhat.com>