skyring-eventing: Modified the event tags to start with skyring.
Modified the event tags to start with skyring. For any event in salt-
master event bus to be listened by skyring it has to have a tag
starting with skyring. We are enforcing this because, with this when
a new storage provider comes to be handled by skyring we need not have
to disrupt the running service(if we were to modify salt config to
handle events with new tag, we will have to restart salt master). So
with this approach any events comming to salt-master event bus with
tag starting with skyring will be listened without disruption.

Change-Id: Ie6bb7224409a919875b731db34045aba410b1f5b
Signed-off-by: nnDarshan <darshan.n.2024@gmail.com>