backend: Corrected to return errors properly from salt.go
Earlier a local variable with same name "err" was getting
created and actual error was not getting returned to the
caller function.

Due to this, in case of some severe errors/exceptions in
invoked python code, the error was not getting propagated
well to the caller functions.

Corrected to explicitly identify the local error instance
and return properly with error instance and other parameters.

Change-Id: I7b5afa41d45c1b88b75c3b4cc13b089d2ab9193f
Signed-off-by: Shubhendu Tripathi <shtripat@redhat.com>