Eventing: Added events to notify node status.
This patch adds events to notify node status. This depends
upon the salt event "salt/presence/change", which continuously
checks if it has connectivity with the minions, if connectivity
is lost or gained, an event is generated. Thus if a node goes
down we can rely on this event to know about it. We will be
listening to this event. whenever we get this event we push
it to skyring.

Change-Id: Iba39d94d035c67fea82896a22d49d87d3a5b76fb
Signed-off-by: nnDarshan <darshan.n.2024@gmail.com>