Eventing: Fix routing provider events when event does not have clusterId.
Fixed routing of provider specific events when an event does not belong
to any cluster (if any provider specific events are recieved even before
cluster creation). Now it checks if a valid provider is returned for the
event, if yes the respective handler is invoked, else it is logged with
proper message.

Change-Id: I531cebed3a2e8e638a0951e285e70913bd5b1ea0
Signed-off-by: nnDarshan <darshan.n.2024@gmail.com>