monitoring : Fix for cluster statistics push to db
When bigfin attempts to call skyring's module, since
the call is in context of bigfin, the configuration used
is also in the context of bigfin. So, it becomes necessary to
pass the required configuration values from bigfin.This
patch alters the framework to accept the values of required
configurations from the provider along with the values to
push to db.

Change-Id: I4d373f451fe619aae4e22f4a2306f0b42c7739a3
Signed-off-by: anmolbabu <anmolbudugutta@gmail.com>