syncDisks: Patch to fix the sync disk functionality.
This patch fixes the sync disk functionality. Earlier this
would assign new uuid to the disks even if they were already
available in DB. Now with this fix it checks if disk is present
or not, only if disk is not present in DB it assigns uuid for it
and updates it in DB. Also if some disk is removed in backend,
that disk will be removed from DB as part of sync.

Change-Id: Ia79adf6d11988ffa9dffe8380bcf66eb3f097586
Signed-off-by: nnDarshan <darshan.n.2024@gmail.com>