AutoExpandCluster: Automatically expand cluster on disk addition.
This patch implements automatic expansation of cluster if a new
disk is added to any of its nodes. First when a new disk is detected
in a node the disk detatils are synced. Later based on if the node
is part of a cluster or not it expands the cluster automatically.

Change-Id: Ie57a5b49db695d8a1dbabb7aae69bd3a9aa09492
Signed-off-by: nnDarshan <darshan.n.2024@gmail.com>