Event: Enhanced the event listing API.
This patch adds filtering capability to events api, events
can be filtered based on severity, ack-state(acked or not),
node, cluster. And pagination support is also added
to the api.
eg: GET request
/api/v1/events?severity=<severity>&acked=<true/false>&
         node-name=<node name>&cluster-name=<cluster name>
         &page-size=<page size>&page-no=<page no>

This patch also adds a new Api to query the details of a
single event based on the event id:
eg: GET request
/api/vi/event/<event-id>

Change-Id: Iab2298296329face95b305b0b34f1225a707d65d
Signed-off-by: nnDarshan <darshan.n.2024@gmail.com>