AlertCorrelation: Clear alerts on recovery events.
This patch adds automatic clearing of alerts, when
corresponding recovery events are received by skyring.
The eventId of the recovery event will be added in
the alert that is cleared. Even if the alert is already
acked the info related to recovery event will be added
to the alert.

Whereas if an alert is automatically acked by backend(
recovery event) then such events cant be acked by the
users.

Change-Id: I89484c996e812d9a1168966216648704f46eea77
Signed-off-by: nnDarshan <darshan.n.2024@gmail.com>