monitoring: Add iops and disk further aggregation
Earlier iops were averaged out across nodes for reads
and writes separately.And the plan was to show writes
in dashboard but now we sum up reads and writes and
intend to show this in dashboard. Same is the case in
network/interface this patch sums up rx and tx.

Change-Id: Iebfa2941b8ac1244f2cebb68208b6ac59d208380
Signed-off-by: anmolbabu <anmolbudugutta@gmail.com>