Update patch set 3
Patch Set 3:

(1 comment)

So here in this patch we are looking for processes ceph-osd and ceph-mon for figure out what storage daemons are running on the node.

What would happen when we get gluster support? This function would start reporting counts as zero.

I feel this daemon's running count should be patch in bigfin side and invoked from skyring core.

Patch-set: 3
Label: Verified=0