monitoring : Get instant stat from collectd - Part 1
As part of this patch instant stat is fetched from
collectd. Following are the statistics that are
implemented using this approach as of now:
1. CPU
2. Memory
3. Swap.
Aggregations of Network and Disk iops, etc... are
currently done from skyring. And they are intended
to be aggregated from collectd. The plan is to take
as much aggregations as possible from skyring to
collectd as part of subsequent patches with this motive.

Change-Id: I5e4bc8f7bf21799e67aaaffb36f8fced349ef4d9
Signed-off-by: anmolbabu <anmolbudugutta@gmail.com>