Url-Redirect: Handles redirection of skyring and graphite web.
This patch handles redirection of skyring and graphite-web on
skyring server. The current behaviour is as follows:

1. Skyring service: By default ssl will be enabled for skyring
service. All the following URLs will redirect to the skyring
login page (ssl enabled https://<host-name>:10443/#/).
* https://<hostname>/skyring
* http://<hostname>/skyring
* <hostname>/skyring

2. Graphite-web: This service will be now running in port 10080
and to access this user can use the following URLs:
* http://<hostname>/graphite-web
* <hostname>/graphite-web

Note: If just hostname is provided then  neither of these services
will be displayed, instead the default index.html on a rhel machine
is displayed.

Change-Id: I89d443fdaa91245a3c444fe1741be78a425b6402
Signed-off-by: Darshan N <darshan.n.2024@gmail.com>