Eventing: checks if node was actually rebooted after gaining contact.
This patch checks if the node was really rebooted, after we get an
event for node gaining contact. This check is done by reading the
value of uptime. If system uptime is less than 5min(300 sec) then
we consider it as rebooted. If rebooted the we have to perform a
sync of the disk information as, they may change after reboot.

Change-Id: I235d373dec28ee5d2a3e611b8d181e74c953ce17
Signed-off-by: Darshan N <darshan.n.2024@gmail.com>