Update patch set 2
Patch Set 2:

Now, We have new UX design for 'Most used storage profile' in main and cluster dashboard , which is recently updated by matt..

You can see here -  (in right top of the page , we have 'Most used storage profile' section)

https://trello-attachments.s3.amazonaws.com/5744136bf5b9ff95558cbaa4/1279x810/c7d4455793ad67eef1769213a9f87147/Dashboard_%28Alternate_536h%29.png 

Here , we are using angular-patternfly graph for rendering this bar chart. and this chart will have different color based on full or near full .

but we don't have any option to set flag for it in graph , so that why we need threshold value for full and near full. so that we can set in graph and depend on that value color of graph will change .

Patch-set: 2