Update patch set 2
Patch Set 2:

Hi Christopher,
As per current design, the application will start only after a successful connection with the database. Application fetches the credentials from a configuration file for an authentication. And the application will not start if the authentication fails or user not found.
The users can be created in the mongodb only after the mongodb service start. Currently mongodb is not friendly for the rpm spec file to execute any mongo shell script to create an user immediately because of the service start delay in either chroot or non-chroot environment. 

So please let this patch get in. We will enhance this part later. I have sent a separate patch (https://review.gerrithub.io/#/c/279558/) for setting an unique random password.

Patch-set: 2
Label: Code-Review=0