skyring-setup: TLS cert usability improvements
Prior to this change, the user had to specify several values to
skyring-setup when generating an SSL certificate, such as a key
passphrase, the various components for the subject field, etc.

With this change, skyring-setup will not prompt the user for anything
when generating the self-signed TLS certificate.

skyring-setup will prompt the user for the hostname first, and then use
that hostname value when constructing the certificate's subject field.

Use an RSA key size of 4096 bits, rather than 1024.

Use an expiration time ten years in the future, rather than a single
year. (This matches what Ceph's calamari application does.)

Change-Id: Ife326b624dc7acbb11a0892c18d167cadf05c30a
Signed-off-by: Ken Dreyer <kdreyer@redhat.com>