Eventing: Dismiss all alerts when an entity is deleted.
This patch adds dismissing of all the events for an entity,
when that entity is removed. To achieve this have modified
the logAuditEvent() utility, and now the caller has to pass
a boolean value indicating if the audit is for a Delete
operation.
In this patch deletes for following entities have been taken
care:
    * cluster
    * node
    * storage
    * block_device

Change-Id: I28423710335b17b09d8671cebc5c894ca7a3ba8e
Signed-off-by: Darshan N <darshan.n.2024@gmail.com>