expand_cluster : Fix cluster expansion case monitoring configuration
Earlier the monitoring configuration was updated using the monitoring
configurations from the cluster. But the cluster contains monitoring
configurations of even the logical resources which are not monitored
using collectd. So, at the time of configuring collectd filter out and
pass only plugins for those resources that are monitored using collectd.

Change-Id: Idf44b4b3c5ddac6cda3c30cc51b80bac77eda5e1
Signed-off-by: anmolbabu <anmolbudugutta@gmail.com>