monitoring : Unmanage/Manage cluster scheduling
This patch removes the monitoring schedule on unmanaging the cluster
and schedules the cluster monitoring on managing back the cluster.
Earlier the schedule was removed on forget cluster and created during
import and create clusters.But now, since the schedule is removed on
unmanage and unmanage is a condition precedent to forget cluster, the
schedule removal is made to happen on unmanage cluster.

Change-Id: Iba1fc9d239db1ec468f4f41cb16959578f90ade7
Signed-off-by: anmolbabu <anmolbudugutta@gmail.com>