skyring.spec: do not label port 8181 for selinux
The ceph-installer application listens on this port, and skyring does
not listen on it. It should not be labelled as skyring_custom_port_t.

When this port is labeled as skyring_custom_port_t and ceph-installer
runs confined in the ceph_installer_t context, SELinux denies
ceph-installer from using this port, like so:

  type=AVC msg=audit(1470235957.124:116): avc:  denied  { name_bind } for
  pid=4957 comm="gunicorn_pecan" src=8181
  scontext=system_u:system_r:ceph_installer_t:s0
  tcontext=system_u:object_r:skyring_custom_port_t:s0 tclass=tcp_socket

Remove the labeling from skyring so ceph-installer can run confined.

Change-Id: I61a7e854fd18cfe76f0c3641d2b62dab7dc3b8f7
Signed-off-by: Ken Dreyer <kdreyer@redhat.com>